move offline reconstruction to apps
