+++ examples/cpp/CMakeLists.txt
@@ -60,7 +60,6 @@ open3d_add_example(SLACIntegrate)
 open3d_add_example(TIntegrateRGBD)
 open3d_add_example(TOdometryRGBD)
 open3d_add_example(TriangleMesh)
-open3d_add_example(LegacyReconstruction Open3D::3rdparty_jsoncpp Open3D::3rdparty_openmp)
 open3d_add_example(TrimMeshBasedOnPointCloud Open3D::3rdparty_openmp)
 open3d_add_example(ViewDistances)
 open3d_add_example(ViewPCDMatch)
